Refactor constraints to sourceConstraints
- Renamed constants and updated related parser methods to reflect the new terminology for source constraints.
- Introduced new `source_constraints_parser.py` and `source_constraints_statement.py` files to handle the parsing and representation of source constraints.
- Updated existing tasks and tests to accommodate the new sourceConstraints structure, ensuring backward compatibility with previous XML configurations.
- Adjusted functional tests to validate the new sourceConstraints implementation in various scenarios.

This commit enhances the clarity and consistency of the constraints handling within the DATAMIMIC framework.
