+++ datamimic_ce/constants/element_constants.py
@@ -25,5 +25,5 @@ EL_GENERATOR = "generator"
 EL_CONDITION = "condition"
 EL_ELSE_IF = "else-if"
 EL_ELSE = "else"
-EL_CONSTRAINTS = "constraints"
+EL_SOURCE_CONSTRAINTS = "sourceConstraints"
 EL_RULE = "rule"
