+++ datamimic_ce/parsers/source_constraints_parser.py
@@ -8,10 +8,10 @@
 from pathlib import Path
 from xml.etree.ElementTree import Element
 
-from datamimic_ce.constants.element_constants import EL_CONSTRAINTS
+from datamimic_ce.constants.element_constants import EL_SOURCE_CONSTRAINTS
 from datamimic_ce.parsers.statement_parser import StatementParser
 from datamimic_ce.statements.composite_statement import CompositeStatement
-from datamimic_ce.statements.constraints_statement import ConstraintsStatement
+from datamimic_ce.statements.source_constraints_statement import ConstraintsStatement
 from datamimic_ce.utils.base_class_factory_util import BaseClassFactoryUtil
 
 
@@ -25,13 +25,13 @@ class ConstraintsParser(StatementParser):
         super().__init__(
             element,
             properties,
-            valid_element_tag=EL_CONSTRAINTS,
+            valid_element_tag=EL_SOURCE_CONSTRAINTS,
             class_factory_util=class_factory_util,
         )
 
     def parse(self, descriptor_dir: Path, parent_stmt: CompositeStatement) -> ConstraintsStatement:
         """
-        Parse element "constraints" into ConstraintsStatement
+        Parse element "sourceConstraints" into ConstraintsStatement
         :return:
         """
         constraints_stmt = ConstraintsStatement(parent_stmt)
