+++ tests_ce/functional_tests/test_constraints/test_constraints.py
@@ -125,7 +125,7 @@ class TestConstraints(TestCase):
             engine.test_with_timer()
             assert False
         except SyntaxError as e:
-            assert "Only once <constraints> allow in per <generate>" in e.msg
+            assert "Only once <sourceConstraints> allow in per <generate>" in e.msg
 
     def test_constraints_nested_data(self):
         engine = DataMimicTest(
