+++ tests_ce/functional_tests/test_constraints/test_constraints_if_rule_empty.xml
@@ -1,9 +1,9 @@
-<setup  multiprocessing="1" numProcess="2">
-    <generate name="synthetic_customers" count="10000" pageSize="1000"
-              source="script/person_data.json" cyclic="True">
-        <key name="id" generator="IncrementGenerator"/>
-        <constraints>
-          <rule if="" then="risk_profile == 'High'"/>
-        </constraints>
-    </generate>
+<setup multiprocessing="1" numProcess="2">
+  <generate name="synthetic_customers" count="10000" pageSize="1000"
+    source="script/person_data.json" cyclic="True">
+    <key name="id" generator="IncrementGenerator" />
+    <sourceConstraints>
+      <rule if="" then="risk_profile == 'High'" />
+    </sourceConstraints>
+  </generate>
 </setup>
