+++ tests_ce/functional_tests/test_constraints/test_constraints_non_cyclic.xml
@@ -1,17 +1,17 @@
-<setup  multiprocessing="1" numProcess="2">
+<setup multiprocessing="1" numProcess="2">
 
-    <generate name="original_customers" count="100" pageSize="50"
-              source="script/person_data.json" target="ConsoleExporter">
-        <key name="id" generator="IncrementGenerator"/>
-    </generate>
+  <generate name="original_customers" count="100" pageSize="50"
+    source="script/person_data.json" target="ConsoleExporter">
+    <key name="id" generator="IncrementGenerator" />
+  </generate>
 
-    <generate name="constraints_customers" count="100" pageSize="50"
-              source="script/person_data.json" target="ConsoleExporter">
-        <key name="id" generator="IncrementGenerator"/>
-        <constraints>
-          <rule if="credit_score &lt; 600" then="risk_profile == 'High'"/>
-          <rule if="credit_score &gt;= 600 and credit_score &lt; 750" then="risk_profile == 'Medium'"/>
-          <rule if="credit_score &gt;= 750" then="risk_profile == 'Low'"/>
-        </constraints>
-    </generate>
+  <generate name="constraints_customers" count="100" pageSize="50"
+    source="script/person_data.json" target="ConsoleExporter">
+    <key name="id" generator="IncrementGenerator" />
+    <sourceConstraints>
+      <rule if="credit_score &lt; 600" then="risk_profile == 'High'" />
+      <rule if="credit_score &gt;= 600 and credit_score &lt; 750" then="risk_profile == 'Medium'" />
+      <rule if="credit_score &gt;= 750" then="risk_profile == 'Low'" />
+    </sourceConstraints>
+  </generate>
 </setup>
