+++ tests_ce/functional_tests/test_constraints/test_nestedkey_constraints.xml
@@ -1,21 +1,21 @@
-<setup  multiprocessing="1" numProcess="2">
-    <generate name="container" count="1">
-        <nestedKey name="cyclic_true" source="script/person_data.json"
-                   type="list" count="1000" cyclic="True">
-            <constraints>
-              <rule if="credit_score &lt; 600" then="risk_profile == 'High'"/>
-              <rule if="credit_score &gt;= 600 and credit_score &lt; 750" then="risk_profile == 'Medium'"/>
-              <rule if="credit_score &gt;= 750" then="risk_profile == 'Low'"/>
-            </constraints>
-            <key name="id" generator="IncrementGenerator"/>
-        </nestedKey>
-        <nestedKey name="cyclic_false" source="script/person_data.json" type="list">
-            <key name="id" generator="IncrementGenerator"/>
-            <constraints>
-              <rule if="credit_score &lt; 600" then="risk_profile == 'High'"/>
-              <rule if="credit_score &gt;= 600 and credit_score &lt; 750" then="risk_profile == 'Medium'"/>
-              <rule if="credit_score &gt;= 750" then="risk_profile == 'Low'"/>
-            </constraints>
-        </nestedKey>
-    </generate>
+<setup multiprocessing="1" numProcess="2">
+  <generate name="container" count="1">
+    <nestedKey name="cyclic_true" source="script/person_data.json"
+      type="list" count="1000" cyclic="True">
+      <sourceConstraints>
+        <rule if="credit_score &lt; 600" then="risk_profile == 'High'" />
+        <rule if="credit_score &gt;= 600 and credit_score &lt; 750" then="risk_profile == 'Medium'" />
+        <rule if="credit_score &gt;= 750" then="risk_profile == 'Low'" />
+      </sourceConstraints>
+      <key name="id" generator="IncrementGenerator" />
+    </nestedKey>
+    <nestedKey name="cyclic_false" source="script/person_data.json" type="list">
+      <key name="id" generator="IncrementGenerator" />
+      <sourceConstraints>
+        <rule if="credit_score &lt; 600" then="risk_profile == 'High'" />
+        <rule if="credit_score &gt;= 600 and credit_score &lt; 750" then="risk_profile == 'Medium'" />
+        <rule if="credit_score &gt;= 750" then="risk_profile == 'Low'" />
+      </sourceConstraints>
+    </nestedKey>
+  </generate>
 </setup>
