+++ tests_ce/functional_tests/test_constraints/test_two_constraints.xml
@@ -1,15 +1,15 @@
-<setup  multiprocessing="1" numProcess="2">
-    <generate name="synthetic_customers" count="10000" pageSize="1000"
-              source="script/person_data.json" cyclic="True">
-        <key name="id" generator="IncrementGenerator"/>
+<setup multiprocessing="1" numProcess="2">
+  <generate name="synthetic_customers" count="10000" pageSize="1000"
+    source="script/person_data.json" cyclic="True">
+    <key name="id" generator="IncrementGenerator" />
 
-        <constraints>
-          <rule if="credit_score &lt; 600" then="risk_profile == 'High'"/>
-          <rule if="credit_score &gt;= 600 and credit_score &lt; 750" then="risk_profile == 'Medium'"/>
-        </constraints>
+    <sourceConstraints>
+      <rule if="credit_score &lt; 600" then="risk_profile == 'High'" />
+      <rule if="credit_score &gt;= 600 and credit_score &lt; 750" then="risk_profile == 'Medium'" />
+    </sourceConstraints>
 
-        <constraints>
-          <rule if="credit_score &gt;= 750" then="risk_profile == 'Low'"/>
-        </constraints>
-    </generate>
+    <sourceConstraints>
+      <rule if="credit_score &gt;= 750" then="risk_profile == 'Low'" />
+    </sourceConstraints>
+  </generate>
 </setup>
